+++ src/windows/BTSerialPortBinding.cc
@@ -90,12 +90,18 @@ void BTSerialPortBinding::EIO_Write(uv_work_t *req) {
     write_baton_t *data = static_cast<write_baton_t*>(queuedWrite->baton);
 
     BTSerialPortBinding *rfcomm = data->rfcomm;
+    int bytesToSend = data->bufferLength;
+    int bytesSent = 0;
 
     if (rfcomm->s != INVALID_SOCKET) {
-        data->result = send(rfcomm->s, (const char *)data->bufferData, data->bufferLength, 0);
-        if (data->result != data->bufferLength) {
-            sprintf_s(data->errorString, "Writing attempt was unsuccessful");
-        }
+        do{
+            data->result = send(rfcomm->s, (const char *)((char*)data->bufferData+bytesSent), bytesToSend, 0);
+            if(data->result != SOCKET_ERROR)
+            {
+                bytesToSend-=data->result;
+                bytesSent+=data->result;
+            }
+        }while(bytesToSend > 0);
     } else {
         sprintf_s(data->errorString, "Attempting to write to a closed connection");
     }
@@ -168,7 +174,7 @@ void BTSerialPortBinding::EIO_AfterRead(uv_work_t *req) {
 
     read_baton_t *baton = static_cast<read_baton_t *>(req->data);
 
-    TryCatch try_catch;
+    Nan::TryCatch try_catch;
 
     Local<Value> argv[2];
 
@@ -189,7 +195,7 @@ void BTSerialPortBinding::EIO_AfterRead(uv_work_t *req) {
     baton->cb->Call(2, argv);
 
     if (try_catch.HasCaught()) {
-        FatalException(try_catch);
+        Nan::FatalException(try_catch);
     }
 
     baton->rfcomm->Unref();
@@ -234,7 +240,7 @@ NAN_METHOD(BTSerialPortBinding::New) {
 
     String::Utf8Value address(info[0]);
     int channelID = info[1]->Int32Value();
-    if (channelID <= 0) { 
+    if (channelID <= 0) {
         Nan::ThrowTypeError("ChannelID should be a positive int value");
     }